ml/requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
librosa==0.10.2.post1
soundfile==0.12.1
tqdm==4.67.1
PyYAML==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
joblib==1.4.2
